TV Guide: Added a hidden button (id: 666) which can be mapped to the back button on your remote.
This will trigger kodi to show full screen video if a channel is playing in the background, otherwise it will execute the normal back action. I find this preferable to the normal "back" action, since if I'm watching TV and pop into the guide, then I'd rather be taken back to Fullscreen video instead of the Home screen when pressing back.